Add missing search for the redirects
